config and default values adjustments

+++ core/src/main/java/io/questdb/std/FilesFacadeImpl.java
@@ -36,8 +36,8 @@ public class FilesFacadeImpl implements FilesFacade {
     public static final int _16M = 16 * 1024 * 1024;
     private long mapPageSize = 0;
     private final AtomicLong mapCount = new AtomicLong();
-    private long openFileLimit = Integer.MAX_VALUE;
-    private long mapLimit = Integer.MAX_VALUE;
+    private long openFileLimit = Long.MAX_VALUE;
+    private long mapLimit = Long.MAX_VALUE;
 
     @Override
     public long append(long fd, long buf, int len) {
